Actually using the real AutumnLeaf class in the spec.
Giving a definition of AutumnLeaf in the spec to test out the removal of the foo_command method wasn't reopening the class, it was defining it. That explains the trouble I was having with other methods that Bot should've inherited from AutumnLeaf, and it definitely explains why Bot could be initialized without any arguments.